Add sampleset helper
